pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "n8n_factory"
version = "1.7.0"
description = "A factory tool to assemble, optimize, and simulate n8n workflows from templates."
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
  {name = "Ryan Rentfro", email = "ryan@example.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Utilities"
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.0",
    "rich>=10.0",
    "requests>=2.28",
    "watchdog>=3.0",
    "python-dotenv>=1.0",
    "deepdiff>=6.0",
    "tabulate>=0.9"
]

[project.urls]
"Homepage" = "https://github.com/username/n8n-factory"
"Bug Tracker" = "https://github.com/username/n8n-factory/issues"

[project.scripts]
n8n-factory = "n8n_factory.cli:main"

[tool.ruff]
line-length = 120
select = ["E", "F", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
log_cli = true
log_cli_level = "INFO"